+++ lua/cyberdream/extra/init.lua
@@ -7,12 +7,14 @@ M.extras = {
     fish = { extension = "theme", name = "fish" },
     gitui = { extension = "ron", name = "gitui" },
     helix = { extension = "toml", name = "helix" },
+    lazydocker = { extension = "yml", name = "lazydocker" },
     lazygit = { extension = "yml", name = "lazygit" },
     pywal = { extension = "json", name = "pywal" },
     textmate = { extension = "tmTheme", name = "textmate" },
     tilix = { extension = "json", name = "tilix" },
     wezterm = { extension = "lua", name = "wezterm" },
     windowsterminal = { extension = "json", name = "windowsterminal" },
+    zellij = { extension = "kdl", name = "zellij" },
 }
 
 --- Create/overwrite a file in the extras directory.